Fixed SD card mounting.
SD card mounting failed previously because /mnt was already mounted
on /dev/block/mmcblk0p13 (boot partition). It was supposed to be
unmounted after the initial LVM setup.

CM seemed to have used a modified init binary that supported umount
as a command. Since we are using stock init binary, this is no longer
supported.

/system is mounted at this stage, so we can take advantage of busybox
stored in it to access umount.